requirements.txt
Django==4.1
python-dotenv==0.20.0
psycopg2-binary==2.9.3
djangorestframework==3.12.0
drf-yasg==1.21.3
django-cors-headers==3.13.0
django-filter==23.2
celery==5.2.7
httpx==0.23.0
asgiref==3.5.2
redis==4.3.4
requests==2.29.0
pytz==2023.3